+++ Boa.Constrictor.UnitTests/WebDriver/Questions/ExistenceTest.cs
@@ -14,16 +14,13 @@ namespace Boa.Constrictor.UnitTests.WebDriver
         [Test]
         public void TestElementExists()
         {
-            var element = new Mock<IWebElement>();
-            WebDriver.Setup(x => x.FindElements(It.IsAny<By>())).Returns(new List<IWebElement> { element.Object }.AsReadOnly());
-
             Actor.AsksFor(Existence.Of(Locator)).Should().BeTrue();
         }
 
         [Test]
         public void TestElementDoesNotExist()
         {
-            WebDriver.Setup(x => x.FindElements(It.IsAny<By>())).Returns(new List<IWebElement>().AsReadOnly());
+            SetUpFindElementsReturnsEmpty();
 
             Actor.AsksFor(Existence.Of(Locator)).Should().BeFalse();
         }