+++ Boa.Constrictor.UnitTests/WebDriver/Questions/PixelSizeTest.cs
@@ -16,8 +16,6 @@ namespace Boa.Constrictor.UnitTests.WebDriver
         [Test]
         public void TestSize()
         {
-            var element = new Mock<IWebElement>();
-            WebDriver.Setup(x => x.FindElements(It.IsAny<By>())).Returns(new List<IWebElement> { element.Object }.AsReadOnly());
             WebDriver.SetupGet(x => x.FindElement(It.IsAny<By>()).Size).Returns(new Size(100, 200));
 
             var point = Actor.AsksFor(PixelSize.Of(Locator));
@@ -28,7 +26,7 @@ namespace Boa.Constrictor.UnitTests.WebDriver
         [Test]
         public void TestElementDoesNotExist()
         {
-            WebDriver.Setup(x => x.FindElements(It.IsAny<By>())).Returns(new List<IWebElement>().AsReadOnly());
+            SetUpFindElementsReturnsEmpty();
 
             Actor.Invoking(x => x.AsksFor(PixelSize.Of(Locator))).Should().Throw<WaitingException<bool>>();
         }