+++ Boa.Constrictor.UnitTests/WebDriver/Questions/SelectedOptionTextTest.cs
@@ -70,7 +70,7 @@ namespace Boa.Constrictor.UnitTests.WebDriver
         [Test]
         public void TestElementDoesNotExist()
         {
-            WebDriver.Setup(x => x.FindElements(It.IsAny<By>())).Returns(new List<IWebElement>().AsReadOnly());
+            SetUpFindElementsReturnsEmpty();
 
             Actor.Invoking(x => x.AsksFor(SelectedOptionText.Of(Locator))).Should().Throw<WaitingException<bool>>();
         }