+++ CHANGELOG.md
@@ -17,7 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-(None)
+### Added
+
+- Added unit tests for Question that derive from `AbstractWebPropertyQuestion`
 
 
 ## [2.0.0] - 2021-11-02